pyproject.toml
[project]
name = "memory"
version = "0.1.0"
description = "A four-tier memory system for AI agents with meta-cognitive reasoning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Marcus Johansson", email = "marcusjihansson@users.noreply.github.com"},
]
keywords = ["memory", "ai", "langgraph", "dspy", "agent", "vector-database"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "dspy-ai>=2.0.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-postgres>=3.0.0",
    "numpy>=2.3.5",
    "pgvector>=0.4.2",
    "psycopg>=3.3.2",
    "python-dotenv>=1.0.0",
    "redis>=7.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0.0",
    "types-requests>=2.31.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memory"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["memory"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
color_output = true

[[tool.mypy.overrides]]
module = ["redis.*", "requests.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/memory"]
branch = true
omit = ["*/tests/*", "*/docs/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
